pyproject.toml
[project]
name = "demo"
version = "1.0.0"
description = "Selenium test suite with BrowserStack integration"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "selenium>=4.16.0",
    "pytest>=7.4.3",
    "pytest-xdist>=3.5.0",
    "pytest-html>=4.1.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pytest-playwright>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --tb=short"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "smoke: Smoke tests for critical functionality",
    "regression: Full regression test suite", 
    "critical: Critical path tests that must pass",
]
